refactor: Move new components to the root components folder
