+++ src/components/Panels.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as TabsPrimitive from '@radix-ui/react-tabs';
 import { cn } from '@/lib/Tailwind';
-import { Typography } from '../Typography';
+import { Typography } from './Typography';
 
 const Tabs = TabsPrimitive.Root;
